requirements.txt
streamlit
ultralytics
numpy
pillow
opencv-python
